Added clearer functions, and stronger runtime cooperation.
Cleaned up a lot of the code, and began ensuring compliance with the don't write it twice methodology. Updated the Graphics Run method to no longer loop, but to set bools to tell the runtime if it should be called or not. Also revised the sine wave function generator method. It will be updated as I learn more about OpenGL and begin re-implementation of the first sprints OpenAL concepts.